update type to double for confidence score. Geo Scene tested producing similar results using test image.

+++ Assets/ScapeKitPlugin/Scripts/ScapeSessionDetails.cs
@@ -59,7 +59,7 @@ namespace ScapeKitUnity
         public Double heading;
         public ScapeOrientation orientation;
         public Double rawHeightEstimate;
-        public long confidenceScore;
+        public Double confidenceScore;
         public ScapeMeasurementStatus measurementsStatus;
     }
